refactor(transform-accessor): move createAccessorDecorator function to a separate file
